feat: refactored gui elements into new module
